+++ run/idp-sql/src/main/java/com/example/cloudrun/IdpSqlApplication.java
@@ -20,10 +20,10 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class CatsVDogsApplication {
+public class IdpSqlApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(CatsVDogsApplication.class, args);
+		SpringApplication.run(IdpSqlApplication.class, args);
 	}
 
 }
