+++ run/idp-sql/src/main/java/com/example/cloudrun/Vote.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2020 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+ package com.example.cloudrun;
+
+import java.sql.Timestamp;
+
+import edu.umd.cs.findbugs.annotations.NonNull;
+import lombok.Data;
+
+/**
+ * Lombok notation to generate methods: 
+ * toString, equals, hashCode, getters, and setters
+ */
+@Data
+public class Vote {
+    @NonNull
+    private String uid;
+    @NonNull
+    private String candidate;
+    @NonNull
+    private Timestamp timeCast;
+}
