+++ run/idp-sql/src/main/resources/application.properties
@@ -13,10 +13,9 @@
 #  limitations under the License.
 server.port=${PORT:8080}
 
+# [START run_user_auth_sql_connect]
 spring.datasource.password=pw-e2e
 spring.cloud.gcp.sql.database-name=postgres
-# format: [gcp-project-id]:[region]:[instance-name]
 spring.cloud.gcp.sql.instance-connection-name=mindful-silo-287418:us-west1:instance-e2e
 spring.datasource.username=postgres
-# So app starts despite "table already exists" errors.
-# spring.datasource.continue-on-error=true
+# [END run_user_auth_sql_connect]
