+++ run/idp-sql/src/main/resources/static/config.js
@@ -1,0 +1,6 @@
+// [START run_end_user_firebase_config]
+const config = {
+  apiKey: "AIzaSyCpgtMGbv3G6NXw3jxZ43AkdLzX-K2BJ8c",
+  authDomain: "mindful-silo-287418.firebaseapp.com",
+};
+// [END run_end_user_firebase_config]
