+++ run/idp-sql/src/main/resources/static/firebase.js
